tenant_id check vs HTTP basic auth for API calls
This is working fine but sub-optimal because we repeat a lot of logic
already in controllers. Ideally we may use a reference to the controller
and call into it.
